--- tests/e2e/tests/proto_msg_to_json.go
@@ -1,25 +1,0 @@
-// Copyright (C) 2022-2025, Chain4Travel AG. All rights reserved.
-// See the file LICENSE for licensing terms.
-
-package tests
-
-import (
-	"google.golang.org/protobuf/encoding/protojson"
-	"google.golang.org/protobuf/proto"
-)
-
-// Service function to convert the responses into pretty-printed JSON.
-// Only used for debugging and test creation.
-func protoMessageToJSON(tt *Test, message proto.Message) string {
-	// Pretty-print using protojson.MarshalOptions
-	marshaler := protojson.MarshalOptions{
-		Multiline: true,
-		Indent:    "  ",
-	}
-	jsonData, err := marshaler.Marshal(message)
-	if err != nil {
-		tt.logger.Errorf("Error marshalling: %v", err)
-		return ""
-	}
-	return string(jsonData)
-}
